requirements.txt
asgiref==3.10.0
Django==5.2.8
sqlparse==0.5.3
gunicorn==21.2.0
djangorestframework==3.14.0
